+++ frontend/src/components/feedback/favoritecharacter/FavoriteCharacter.tsx
