+++ frontend/src/components/feedback/staricon/StarIcon.css
